Added in sample code for updating
The image will now loop through colours for one of the pixels drawn over
the background image. This shows that it will be possible to draw a
static background and have entities move around on top of it.

The setInterval() function may also be what we're after for a main game
loop, containing the update and render functions.